requirements.txt
pandas==2.2.3
pandas_ta==0.3.14b0
python-telegram-bot==21.7
python_binance==1.0.22
Requests==2.32.3
